reapply inversion after DOM content loaded
Added boolean return type to autoLoad.  If it returns true (indicating inversion was automatically applied to the current page because it was in the autoChange list), then add an event listener for DOMContentLoaded such that inversion is applied once more.  (Note that forceEnable is true on both occasions.)  This better handles scenarios where the page is partially loaded when the first invertColor call is performed, leading to a page presentation that would otherwise be "half-inverted", if not for the second invertColor call.